move sym_link from fileplatform to fileoperations
