+++ idmtools_platform_container/tests/test_container_platform.py
@@ -105,7 +105,6 @@ class TestContainerPlatform(unittest.TestCase):
             self.assertEqual(platform.data_mount, '/home/container_data')
             self.assertEqual(platform.force_start, False)
             self.assertEqual(platform.user_mounts, None)
-            self.assertEqual(platform.sym_link, True)
 
     @patch('docker.from_env')
     def test_start_container_custom_data_mount(self, mock_docker):
