+++ idmtools_platform_slurm/tests/test_slurm_operations.py
@@ -1,236 +1,291 @@
 import os
+import pathlib
+import shutil
 import tempfile
-from functools import partial
-from pathlib import Path
-
 import pytest
-from idmtools.builders import SimulationBuilder
-from idmtools.core import ItemType, EntityStatus
 
+from idmtools import IdmConfigParser
 from idmtools.core.platform_factory import Platform
+from idmtools.entities import Suite
 from idmtools.entities.experiment import Experiment
 from idmtools.entities.simulation import Simulation
-from idmtools_models.python.json_python_task import JSONConfiguredPythonTask
-from idmtools_platform_file.platform_operations.utils import FileExperiment, FileSimulation, add_dummy_suite
-from idmtools_test import COMMON_INPUT_PATH
+from idmtools_platform_slurm.slurm_operations.slurm_operations import SlurmOperations
 from idmtools_test.utils.decorators import linux_only
+
 from idmtools_test.utils.itest_with_persistence import ITestWithPersistence
+from idmtools_test.utils.test_task import TestTask
 
-from idmtools.entities.templated_simulation import TemplatedSimulations
-from idmtools.assets.asset import Asset
-
-setA = partial(JSONConfiguredPythonTask.set_parameter_sweep_callback, param="a")
+job_dir = os.path.expanduser('~')
 
 
+@pytest.mark.smoke
 @pytest.mark.serial
-@linux_only
-class TestSlurmOperations(ITestWithPersistence):
-
-    def create_experiment(self, platform=None):
-        task = JSONConfiguredPythonTask(script_path=os.path.join(COMMON_INPUT_PATH, 'python', 'model1.py'),
-                                        parameters=dict(c='c-value'))
-        task.common_assets.add_asset('input/hello.sh')
-        ts = TemplatedSimulations(base_task=task)
-        builder = SimulationBuilder()
-        builder.add_sweep_definition(setA, range(0, 2))
-        ts.add_builder(builder)
-        exp = Experiment(name=self.case_name, simulations=ts, tags=dict(number_tag=123, KeyOnly=None))
-        suite = add_dummy_suite(exp)
-        suite.run(platform=platform, wait_until_done=False, dry_run=True)
-        return suite, exp
+class TestSlurmPlatform(ITestWithPersistence):
 
     def setUp(self) -> None:
-        self.case_name = os.path.basename(__file__) + '--' + self._testMethodName
-        self.job_directory = 'DEST'
-        self.platform = Platform('SLURM_LOCAL', job_directory=self.job_directory)
-        self.suite, self.exp = self.create_experiment(self.platform)
+        self.platform = Platform('SLURM_LOCAL')
 
-    def test_retrieve_experiment(self):
-        exp = self.platform.get_item(self.exp.uid, ItemType.EXPERIMENT)
-        # Test attributes
-        self.assertEqual(self.exp.uid, exp.uid)
-        self.assertEqual(self.exp.name, exp.name)
+    # Test platform slurm_fields property
+    def test_slurm_platform_fields(self):
+        actual_field_set = self.platform.slurm_fields
+        expected_field_set = {'mem', 'partition', 'time', 'requeue', 'mail_user', 'ntasks', 'modules',
+                              'exclusive', 'mail_type', 'sbatch_custom', 'nodes', 'ntasks_per_core', 'account',
+                              'mem_per_cpu', 'max_running_jobs', 'cpus_per_task', 'constraint', 'mpi_type'}
 
-        # Comps returns tags as string regardless of type
-        self.assertEqual({k: (v or None) for k, v in self.exp.tags.items()}, exp.tags)
+        self.assertEqual(set(expected_field_set), set(actual_field_set))
 
-        # Test the raw retrieval
-        slurm_experiment = self.platform.get_item(self.exp.uid, ItemType.EXPERIMENT, raw=True)
-        self.assertIsInstance(slurm_experiment, FileExperiment)
-        self.assertEqual(str(self.exp.uid), slurm_experiment.uid)
-        self.assertEqual(self.exp.name, slurm_experiment.name)
-        self.assertEqual({k: (v or None) for k, v in self.exp.tags.items()}, slurm_experiment.tags)
+    # Test platform get_slurm_configs with default config
+    def test_slurm_configs_default(self):
+        slurm_configs_dict = self.platform.get_slurm_configs()
+        expected_config_dict = {'mem': None, 'time': None, 'modules': [], 'mail_user': None,
+                                'exclusive': False, 'sbatch_custom': None, 'nodes': None, 'mail_type': None,
+                                'partition': None, 'account': None, 'ntasks_per_core': None, 'requeue': True,
+                                'max_running_jobs': None, 'mem_per_cpu': None, 'ntasks': None, 'cpus_per_task': None,
+                                'constraint': None, 'mpi_type': 'pmi2'}
 
-        # Test retrieving less columns
-        slurm_experiment = self.platform.get_item(self.exp.uid, ItemType.EXPERIMENT, raw=True)
-        self.assertEqual(str(self.exp.uid), slurm_experiment.uid)
-        self.assertEqual(self.exp.name, slurm_experiment.name)
-        self.assertEqual(str(self.exp.uid), slurm_experiment.uid)
+        self.assertEqual(set(slurm_configs_dict), set(expected_config_dict))
 
-    def test_retrieve_simulation(self):
-        base = self.exp.simulations[0]
-        sim = self.platform.get_item(base.uid, ItemType.SIMULATION)
-        # Test attributes
-        self.assertEqual(sim.uid, base.uid)
-        self.assertEqual(sim.name, base.name)
-        self.assertEqual({k: v for k, v in base.tags.items()}, sim.tags)
+    # Test platform get_slurm_configs with user defined configs
+    def test_slurm_configs_from_user_defined(self):
+        IdmConfigParser.clear_instance()
+        platform = Platform("SLURM_TEST", job_directory=".", mode="local", mail_user="test@test.com",
+                            account="test_acct", mail_type="begin", mem_per_cpu=2048)
+        slurm_configs_dict = platform.get_slurm_configs()
+        expected_config_dict = {'account': 'test_acct', 'exclusive': False, 'mail_type': 'begin',
+                                'mail_user': 'test@test.com', 'max_running_jobs': None, 'mem': None,
+                                'mem_per_cpu': 2048, 'modules': [], 'nodes': None, 'ntasks': None,
+                                'ntasks_per_core': None, 'partition': None, 'requeue': True,
+                                'sbatch_custom': None, 'time': None, 'cpus_per_task': None, 'constraint': None,
+                                'mpi_type': 'pmi2'}
 
-        # Test the raw retrieval
-        slurm_simulation = self.platform.get_item(base.uid, ItemType.SIMULATION, raw=True)
-        self.assertIsInstance(slurm_simulation, FileSimulation)
-        self.assertEqual(str(base.uid), slurm_simulation.uid)
-        self.assertEqual(self.case_name, slurm_simulation.name)
-        self.assertEqual({k: v for k, v in base.tags.items()}, slurm_simulation.tags)
+        self.assertEqual(set(slurm_configs_dict), set(expected_config_dict))
 
-    def test_parent(self):
-        parent_exp = self.platform.get_parent(self.exp.simulations[0].uid, ItemType.SIMULATION)
-        self.assertEqual(self.exp.uid, parent_exp.uid)
-        self.assertEqual({k: v or None for k, v in self.exp.tags.items()}, parent_exp.tags)
-        parent_suite = self.platform.get_parent(self.exp.uid, ItemType.EXPERIMENT)
-        self.assertEqual(self.suite.uid, parent_suite.uid)
+        # validate custom default config get override with Platform parameters
+        self.assertEqual(platform.job_directory, os.getcwd())
+        self.assertEqual(platform.mode.name, "LOCAL")
+        self.assertEqual(platform.mode.value, "local")
 
-    # To test get_children with raw=False(default). Children should be idmtools Simulation or Experiment type
-    def test_children(self):
-        with self.subTest('test_get_children_for_experiment'):
-            children = self.platform.get_children(self.exp.uid, ItemType.EXPERIMENT)
-            for child in children:
-                self.assertTrue(isinstance(child, Simulation))
-                # this depends on the config for the op so we need to make this test ensure it uses correct config
-                self.assertTrue(isinstance(child.uid, str))
-            self.assertEqual(len(self.exp.simulations), len(children))
-            for s in self.exp.simulations:
-                self.assertIn(s.uid, [s.uid for s in children])
-            self.assertCountEqual(self.platform.get_children(self.exp.simulations[0].uid, ItemType.SIMULATION), [])
+    # Test LocalSlurmOperations get_directory for suite only case
+    def test_localSlurmOperations_get_directory_suite(self):
+        suite = Suite(name="Suite")
+        slurm_op = SlurmOperations(platform=self.platform)
+        actual_entity_path = slurm_op.get_directory(suite)
+        suite_dir = str(self.platform.get_directory(suite))
+        self.assertEqual(str(actual_entity_path), suite_dir)
 
-        with self.subTest('test_get_children_for_suite'):
-            children = self.platform.get_children(self.suite.uid, ItemType.SUITE)
-            for child in children:
-                self.assertTrue(isinstance(child, Experiment))
-                self.assertTrue(isinstance(child.uid, str))
-            self.assertEqual(len(children), 1)
-            for e in self.suite.experiments:
-                self.assertIn(e.uid, [e.uid for e in children])
+    # Test LocalSlurmOperations get_directory for multiple experiments case
+    def test_localSlurmOperations_get_directory_experiments(self):
+        suite = Suite(name="Suite")
+        exp1 = Experiment(name="Exp1")
+        exp1.parent = suite
+        exp2 = Experiment(name="Exp2")
+        exp2.parent = suite
+        exp3 = Experiment(name="Exp3")
+        exp3.parent = suite
+        exp_dir1 = str(self.platform.get_directory(exp1))
+        suite_dir = str(self.platform.get_directory(suite))
+        slurm_op = SlurmOperations(platform=self.platform)
+        actual_suite_path = slurm_op.get_directory(suite)
+        self.assertEqual(str(actual_suite_path), suite_dir)
 
-    # To test get_children with raw=True. Children should be Slurm type
-    def test_children_with_raw(self):
-        with self.subTest('test_get_children_for_experiment_raw_true'):
-            children = self.platform.get_children(self.exp.uid, ItemType.EXPERIMENT, raw=True)
-            self.assertEqual(len(self.exp.simulations), len(children))
-            for child in children:
-                self.assertTrue(isinstance(child, FileSimulation))
-                self.assertTrue(isinstance(child.uid, str))
-        with self.subTest('test_get_children_for_suite_raw_true'):
-            children = self.platform.get_children(self.suite.uid, ItemType.SUITE, raw=True)
-            self.assertEqual(len(children), 1)
-            for child in children:
-                self.assertTrue(isinstance(child, FileExperiment))
-                self.assertTrue(isinstance(child.uid, str))
+        actual_exp1_path = slurm_op.get_directory(exp1)
+        self.assertEqual(str(actual_exp1_path), exp_dir1)
 
-    def test_experiment_list_assets(self):
-        with self.subTest('test_list_assets'):
-            assets = self.platform._experiments.list_assets(self.exp)
-            self.assertEqual(2, len(assets))
-            self.assertEqual(set([asset.filename for asset in assets]), set(['model1.py', 'hello.sh']))
-            experiment_dir = Path.resolve(self.platform.get_directory(self.exp))
-            expected_assets_path = [str(experiment_dir.joinpath('Assets/model1.py')),
-                                    str(experiment_dir.joinpath('Assets/hello.sh'))]
-            self.assertEqual(set([asset.absolute_path for asset in assets]), set(expected_assets_path))
+    # Test LocalSlurmOperations get_directory for experiment only
+    def test__localSlurmOperations_get_directory_experiment_no_suite(self):
+        exp = Experiment(name="Exp1")
+        slurm_op = SlurmOperations(platform=self.platform)
+        with self.assertRaises(RuntimeError) as ex:
+            actual_entity_path = slurm_op.get_directory(exp)
+        self.assertEqual(ex.exception.args[0], "Experiment missing parent!")
 
-        with self.subTest('test_list_assets_add_exclude'):
-            assets = self.platform._simulations.list_assets(self.exp.simulations[0], exclude='hello.sh')
-            self.assertEqual(4, len(assets))
-            self.assertTrue(["hello.sh" is not asset.filename for asset in assets])
+    # Test LocalSlurmOperations get_directory for simulation and experiment as parent
+    def test_localSlurmOperations_get_directory_simulation(self):
+        suite = Suite(name="Suite")
+        experiment = Experiment(name="Exp1")
+        experiment.parent = suite
+        simulation = Simulation()
+        simulation.parent = experiment
+        slurm_op = SlurmOperations(platform=self.platform)
+        actual_entity_path = slurm_op.get_directory(simulation)
+        sim_dir = str(self.platform.get_directory(simulation))
+        self.assertEqual(str(actual_entity_path), sim_dir)
 
-    def test_simulation_list_assets(self):
-        count = 0
-        for sim in self.exp.simulations:
-            assets = self.platform._simulations.list_assets(sim)
-            self.assertEqual(3, len(assets))
-            self.assertEqual(set([asset.filename for asset in assets]),
-                             set(['model1.py', 'hello.sh', 'config.json']))
-            simulation_dir = Path.resolve(self.platform.get_directory(sim))
-            expected_assets_path = [str(simulation_dir.joinpath('Assets/model1.py')),
-                                    str(simulation_dir.joinpath('Assets/hello.sh')),
-                                    str(simulation_dir.joinpath('config.json'))]
-            self.assertEqual(set([asset.absolute_path for asset in assets]), set(expected_assets_path))
-            count += 1
-        self.assertEqual(count, 2)
+    # Test LocalSlurmOperations get_directory for simulation without parent, it should throw error
+    def test_localSlurmOperations_get_simulation_no_experiment_entity_dir(self):
+        simulation = Simulation(name="Sim1")
+        slurm_op = SlurmOperations(platform=self.platform)
+        with self.assertRaises(RuntimeError) as ex:
+            actual_entity_path = slurm_op.get_directory(simulation)
+        self.assertEqual(ex.exception.args[0], "Simulation missing parent!")
 
-    def test_to_entity(self):
-        slurm_experiment = self.platform.get_item(self.exp.id, item_type=ItemType.EXPERIMENT, raw=True)
-        idm_experiment = self.platform._experiments.to_entity(slurm_experiment)
-        self.assertEqual(slurm_experiment.id, idm_experiment.id)
-        self.assertEqual(slurm_experiment.uid, idm_experiment.id)
-        self.assertEqual(slurm_experiment.name, idm_experiment.name)
-        self.assertEqual(slurm_experiment.tags, idm_experiment.tags)
-        self.assertFalse(hasattr(slurm_experiment, 'simulations'))
+    # Test LocalSlurmOperations mk_directory with suite and without dest
+    def test_localSlurmOperations_make_dir_no_dest_suite(self):
+        suite = Suite(name="Suite")
+        suite_dir = str(self.platform.get_directory(suite))
+        slurm_op = SlurmOperations(platform=self.platform)
+        slurm_op.mk_directory(suite)
+        self.assertTrue(os.path.isdir(suite_dir))
+        # delete dir after test
+        shutil.rmtree(suite_dir)
+        self.assertFalse(os.path.isdir(suite_dir))
 
-        # we only compare asset filenames
-        slurm_experiment_assets = [asset['filename'] for asset in slurm_experiment.assets]
-        idm_experiment_assets = [asset.filename for asset in idm_experiment.assets]
-        self.assertEqual(set(slurm_experiment_assets), set(idm_experiment_assets))
-        self.assertEqual(slurm_experiment.status, 'CREATED')
-        self.assertEqual(idm_experiment.status, EntityStatus.CREATED)
+    # Test LocalSlurmOperations mk_directory with dest
+    def test_localSlurmOperations_make_dir_dest_suite(self):
+        suite = Suite(name="Suite")
+        slurm_op = SlurmOperations(platform=self.platform)
+        dest_dir = tempfile.mkdtemp()
+        slurm_op.mk_directory(suite, dest=dest_dir, exist_ok=True)
+        expected_dir = os.path.join(dest_dir)
+        self.assertTrue(os.path.isdir(expected_dir))
 
-    def test_get_files(self):
-        with self.subTest('test_get_files_for_experiment'):
-            exp_files = self.platform.get_files(self.exp, ['Assets/model1.py', 'Assets/hello.sh'])
-            with open(os.path.join(COMMON_INPUT_PATH, 'python', 'model1.py'), 'rb') as m:
-                self.assertEqual(exp_files[self.exp.simulations[0].uid]['Assets/model1.py'], m.read())
-            with open('input/hello.sh', 'rb') as m:
-                self.assertEqual(exp_files[self.exp.simulations[0].uid]['Assets/hello.sh'], m.read())
+        # make sure we do not create suite id folder under dest folder in this case
+        self.assertFalse(os.path.isdir(os.path.join(dest_dir, suite.id)))
 
-        with self.subTest('test_get_files_for_simulation'):
-            # get_file from one of simulations
-            output_dir = tempfile.TemporaryDirectory().name  # save files to temp dir
-            files_needed = ["config.json", 'Assets/model1.py', 'Assets/hello.sh']
-            files_retrieved = self.platform.get_files(item=self.exp.simulations[0], files=files_needed,
-                                                      output=output_dir)
-            with open(os.path.join(output_dir, self.exp.simulations[0].id, 'Assets', 'model1.py'), 'rb') as m:
-                self.assertEqual(files_retrieved['Assets/model1.py'], m.read())
-            with open(os.path.join(output_dir, self.exp.simulations[0].id, 'Assets', 'hello.sh'), 'rb') as m:
-                self.assertEqual(files_retrieved['Assets/hello.sh'], m.read())
-            with open(os.path.join(output_dir, self.exp.simulations[0].id, 'config.json'), 'rb') as m:
-                self.assertEqual(files_retrieved['config.json'], m.read())
+    # Test LocalSlurmOperations mk_directory without entity and dest
+    def test_localSlurmOperations_make_dir_no_params(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        with self.assertRaises(RuntimeError) as ex:
+            slurm_op.mk_directory()
+        self.assertEqual(ex.exception.args[0], "Only support Suite/Experiment/Simulation or not None dest.")
 
-        with self.subTest('test_get_files_for_experiment_no_existing'):
-            with self.assertRaises(RuntimeError) as context:
-                self.platform.get_files(self.exp, ['Assets/no_existing.txt'])
-            self.assertTrue(
-                    "Couldn't find asset for path 'Assets/no_existing.txt'." in str(context.exception.args[0]))
+    # Test LocalSlurmOperations mk_directory for simulation
+    def test_localSlurmOperations_make_dir_dest_simulations(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        suite = Suite(name="Suite")
+        experiment = Experiment(name="Exp1")
+        experiment.parent = suite
+        simulation1 = Simulation(name="Sim1")  # create uniq simulation
+        simulation2 = Simulation(name="Sim2")
+        simulation1.parent = experiment
+        simulation2.parent = experiment
+        slurm_op.mk_directory(simulation1)
+        slurm_op.mk_directory(simulation2)
+        suite_dir = str(self.platform.get_directory(suite))
+        sim_dir1 = str(self.platform.get_directory(simulation1))
+        sim_dir2 = str(self.platform.get_directory(simulation2))
+        self.assertTrue(os.path.isdir(sim_dir1))
+        self.assertTrue(os.path.isdir(sim_dir2))
+        # delete dir after test
+        shutil.rmtree(suite_dir)
+        self.assertFalse(os.path.isdir(sim_dir1))
+        self.assertFalse(os.path.isdir(sim_dir2))
 
-    def test_get_files_by_id(self):
-        with self.subTest('test_get_files_by_id_for_experiment'):
-            exp_files = self.platform.get_files_by_id(self.exp.id, ItemType.EXPERIMENT,
-                                                      ['Assets/model1.py', 'Assets/hello.sh'])
-            with open(os.path.join(COMMON_INPUT_PATH, 'python', 'model1.py'), 'rb') as m:
-                self.assertEqual(exp_files[self.exp.simulations[0].uid]['Assets/model1.py'], m.read())
-            with open('input/hello.sh', 'rb') as m:
-                self.assertEqual(exp_files[self.exp.simulations[0].uid]['Assets/hello.sh'], m.read())
+    # Test LocalSlurmOperations create_batch_file for experiment
+    def test_localSlurmOperations_create_batch_file_experiment(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        suite = Suite(name="Suite")
+        experiment = Experiment(name="Exp1")
+        experiment.parent = suite
+        exp_dir = str(self.platform.get_directory(experiment))
+        suite_dir = str(self.platform.get_directory(suite))
+        slurm_op.mk_directory(experiment)
+        slurm_op.create_batch_file(experiment)
+        # verify batch file locally
+        job_path = os.path.join(exp_dir, "sbatch.sh")
+        self.assertTrue(os.path.exists(job_path))
+        # TODO validation sbatch.sh content
+        with open(job_path) as f:
+            contents = f.read()
+        # check srun run_simulation.sh in sbatch.sh file
+        self.assertIn(
+            "bash run_simulation.sh",
+            contents)
+        # clean up suite folder
+        shutil.rmtree(suite_dir)
+        self.assertFalse(os.path.exists(job_path))
 
-        with self.subTest('test_get_files_by_id_for_simulation'):
-            # get_file from one of simulations
-            output_dir = tempfile.TemporaryDirectory().name  # save files to temp dir
-            files_needed = ["config.json", 'Assets/model1.py', 'Assets/hello.sh']
-            files_retrieved = self.platform.get_files_by_id(self.exp.simulations[0].id, ItemType.SIMULATION,
-                                                            files=files_needed, output=output_dir)
-            with open(os.path.join(output_dir, self.exp.simulations[0].id, 'Assets', 'model1.py'), 'rb') as m:
-                self.assertEqual(files_retrieved['Assets/model1.py'], m.read())
-            with open(os.path.join(output_dir, self.exp.simulations[0].id, 'Assets', 'hello.sh'), 'rb') as m:
-                self.assertEqual(files_retrieved['Assets/hello.sh'], m.read())
-            with open(os.path.join(output_dir, self.exp.simulations[0].id, 'config.json'), 'rb') as m:
-                self.assertEqual(files_retrieved['config.json'], m.read())
+    # Test LocalSlurmOperations create_batch_file for simulation
+    def test_localSlurmOperations_create_batch_file_simulation(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        suite = Suite(name="Suite")
+        experiment = Experiment(name="Exp1")
+        experiment.parent = suite
+        simulation = Simulation(task=TestTask())
+        simulation.parent = experiment
+        suite_dir = str(self.platform.get_directory(suite))
+        simulation_dir = str(self.platform.get_directory(simulation))
+        slurm_op.mk_directory(simulation)
+        slurm_op.create_batch_file(simulation)
+        # verify batch file
+        job_path = os.path.join(simulation_dir, "_run.sh")
+        self.assertTrue(os.path.exists(job_path))
+        # TODO validation _run.sh content
+        # clean up suite folder
+        shutil.rmtree(suite_dir)
+        self.assertFalse(os.path.exists(job_path))
 
-    def test_get_files_from_dynamic_created(self):
-        task = JSONConfiguredPythonTask(script_path=os.path.join(COMMON_INPUT_PATH, 'python', 'model1.py'),
-                                        parameters=dict(c='c-value'))
-        task.common_assets.add_asset(Asset(content="test", filename="test.txt"))
-        exp = Experiment.from_task(task, name=self.case_name)
-        suite = add_dummy_suite(exp)
-        platform = Platform('SLURM_LOCAL', job_directory="test")
-        suite.run(platform=platform, wait_until_done=False, dry_run=True)
-        my_exp = platform.get_item(exp.id, ItemType.EXPERIMENT)
-        self.assertEqual(my_exp.id, exp.id)
-        assets = my_exp.assets
-        self.assertEqual(2, len(assets))
-        self.assertEqual(set([asset.filename for asset in assets]), set(['model1.py', 'test.txt']))
+    # Test LocalSlurmOperations create_batch_file with simulation and item_path
+    def test_localSlurmOperations_create_batch_file_simulation_and_item_path(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        suite = Suite(name="Suite")
+        experiment = Experiment(name="Exp1")
+        experiment.parent = suite
+        simulation = Simulation(task=TestTask())
+        simulation.parent = experiment
+        temp_path = tempfile.mkdtemp()
+        slurm_op.mk_directory(simulation, dest=temp_path)
+        slurm_op.create_batch_file(simulation, item_path=temp_path)
+        # verify batch file
+        job_path = os.path.join(temp_path, "_run.sh")
+        self.assertTrue(os.path.exists(job_path))
+        # TODO validation _run.sh content
+
+    # Test LocalSlurmOperations create_batch_file with suite and item_path. this case will throw error
+    def test_localSlurmOperations_create_batch_file_simulation_and_item_path(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        suite = Suite(name="Suite")
+        temp_path = tempfile.mkdtemp()
+        slurm_op.mk_directory(suite, dest=temp_path, exist_ok=True)
+        with self.assertRaises(NotImplementedError) as ex:
+            slurm_op.create_batch_file(suite, item_path=temp_path)
+        self.assertEqual(ex.exception.args[0], "Suite is not supported for batch creation.")
+
+    @linux_only
+    def test_localSlurmOperations_link_file(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        temp_source_path = tempfile.mkdtemp()
+        temp_dest_path = tempfile.mkdtemp()
+        target_file = "test.txt"
+        with open(os.path.join(temp_source_path, target_file), 'w') as fp:
+            fp.write("this is test file")
+
+        # First test with filepath for target and link
+        slurm_op.link_file(target=os.path.join(temp_source_path, target_file),
+                        link=os.path.join(temp_dest_path, target_file))
+        self.assertTrue(os.path.exists(os.path.join(temp_dest_path, target_file)))
+        with open(os.path.join(temp_dest_path, target_file), 'r') as fpr:
+            contents = fpr.read()
+        self.assertEqual(contents, "this is test file")
+
+        # Second test with filepath as string for target and link path as string
+        temp_source_path = tempfile.mkdtemp()
+        temp_dest_path = tempfile.mkdtemp()
+        with open(os.path.join(temp_source_path, target_file), 'w') as fp:
+            fp.write("this is test file1")
+        slurm_op.link_file(target=temp_source_path + "/" + target_file, link=temp_dest_path + "/" + target_file)
+        self.assertTrue(os.path.exists(os.path.join(temp_dest_path, target_file)))
+        with open(os.path.join(temp_dest_path, target_file), 'r') as fpr:
+            contents = fpr.read()
+        self.assertEqual(contents, "this is test file1")
+
+    @linux_only
+    def test_localSlurmOperations_link_dir(self):
+        slurm_op = SlurmOperations(platform=self.platform)
+        temp_source_path = tempfile.mkdtemp()
+        dest_path = "DEST_TEST"
+        target_file = "test.txt"
+        with open(os.path.join(temp_source_path, target_file), 'w') as fp:
+            fp.write("this is test file")
+        slurm_op.link_dir(target=temp_source_path, link=dest_path)
+        self.assertTrue(os.path.exists(os.path.join(dest_path, target_file)))
+        with open(os.path.join(dest_path, target_file), 'r') as fpr:
+            contents = fpr.read()
+        self.assertEqual(contents, "this is test file")
+
+        # Clean up dest_path dir
+        for path in pathlib.Path(dest_path).glob("**/*"):
+            if path.is_file():
+                path.unlink()
+        if os.path.isdir(dest_path):
+            pathlib.Path(dest_path).unlink()
